+++ common/eve_hal/EVE_CoCmd.c
@@ -294,7 +294,7 @@ void EVE_CoCmd_endList(EVE_HalContext *phost)
 
 bool EVE_CoCmd_getMatrix(EVE_HalContext *phost, int32_t *m)
 {
-	uint16_t resAddr;
+	uint32_t resAddr;
 	int i;
 
 #if EVE_CMD_HOOKS
